moved the task api tests into individual files
stops conflicts when we all work on tasks
