+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.API.Contracts/Project/Tasks/TrustTask.cs
@@ -11,7 +11,6 @@ namespace Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Tasks
     
     public enum TrustType
     {
-        [Description("")]
         NotSet,
         [Description("SAT (single academy trust)")]
         SingleAcademyTrust,
