+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects/Pages/Project/Tasks/Trust/ViewTrustTask.cshtml
@@ -23,8 +23,8 @@
 
         <govuk-summary-list>
             <govuk-summary-item label="TRN (trust reference number)" asp-for="@trustTask.TRN" href="@editlink" />
-            <govuk-summary-item label="Trust name" asp-for="@trustTask.TrustName" href="@editlink" render-link="false" />
-            <govuk-summary-item label="Trust type" asp-for="@trustTask.TrustType" href="@editlink" render-link="false" />
+            <govuk-summary-item label="Trust name" asp-for="@trustTask.TrustName" render-link="false" />
+            <govuk-summary-item label="Trust type" asp-for="@trustTask.TrustType" render-link="false" />
         </govuk-summary-list>
 
         <form method="post">
